+++ sandbox/notification-socket-example/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>2.0.1 (2021-05-19)</small>
+
+**Note:** Version bump only for package @sourceloop/notification-socket-example
+
+
+
+
+
 ## 2.0.0 (2021-05-12)
 
 **Note:** Version bump only for package @sourceloop/notification-socket-example